# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorDetail": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					}
				]
			},
			"ErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"$ref": "#/components/schemas/ErrorDetail"
					},
					"timestamp": {
						"type": "string"
					},
					"developer": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
			"IcommonImage": {
				"properties": {
					"url": {
						"type": "string"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"url",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidUnit": {
				"type": "string",
				"enum": [
					"piece",
					"kg",
					"gm",
					"litre",
					"ml",
					"pack",
					"dozen"
				]
			},
			"ProductType": {
				"type": "string",
				"enum": [
					"Food",
					"Electronics",
					"Apparel",
					"General"
				]
			},
			"IFoodProductDetails": {
				"properties": {
					"fssaiLicenceNumber": {
						"type": "string"
					},
					"isVegetarian": {
						"type": "boolean"
					},
					"shelfLife": {
						"type": "string"
					},
					"keyFeatures": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IElectronicsProductDetails": {
				"properties": {
					"modelNumber": {
						"type": "string"
					},
					"warranty": {
						"type": "string"
					},
					"specifications": {
						"items": {
							"properties": {
								"value": {
									"type": "string"
								},
								"key": {
									"type": "string"
								}
							},
							"required": [
								"value",
								"key"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IApparelProductDetails": {
				"properties": {
					"size": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"color": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"material": {
						"type": "string"
					},
					"careInstructions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IGeneralProductDetails": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"ProductDetails": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IFoodProductDetails"
					},
					{
						"$ref": "#/components/schemas/IElectronicsProductDetails"
					},
					{
						"$ref": "#/components/schemas/IApparelProductDetails"
					},
					{
						"$ref": "#/components/schemas/IGeneralProductDetails"
					}
				]
			},
...
			"CustomerProductResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"mrp": {
						"type": "number",
						"format": "double"
					},
					"sellingPrice": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"sku": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/IcommonImage"
						},
						"type": "array"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"inStock": {
						"type": "boolean"
					},
					"brand": {
						"type": "string"
					},
					"unit": {
						"$ref": "#/components/schemas/ValidUnit"
					},
					"productType": {
						"$ref": "#/components/schemas/ProductType"
					},
					"productDetails": {
						"$ref": "#/components/schemas/ProductDetails"
					},
					"averageRating": {
						"type": "number",
						"format": "double"
					},
					"reviewCount": {
						"type": "number",
						"format": "double"
					},
					"discountPercentage": {
						"type": "number",
						"format": "double"
					},
					"savings": {
						"type": "number",
						"format": "double"
					},
					"hasDiscount": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"isActive",
					"_id",
					"slug",
					"mrp",
					"sellingPrice",
					"category",
					"sku",
					"quantity",
					"isFeatured",
					"inStock",
					"unit",
					"productType",
					"productDetails",
					"averageRating",
					"reviewCount",
					"discountPercentage",
					"savings",
					"hasDiscount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_CustomerProductResponse_": {
				"properties": {
					"docs": {
						"items": {
							"$ref": "#/components/schemas/CustomerProductResponse"
						},
						"type": "array"
					},
					"totalDocs": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPrevPage": {
						"type": "boolean"
					},
					"nextPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"prevPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"docs",
					"totalDocs",
					"limit",
					"page",
					"totalPages",
					"hasNextPage",
					"hasPrevPage",
					"nextPage",
					"prevPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_PaginatedResponse_CustomerProductResponse__": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/PaginatedResponse_CustomerProductResponse_"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IProduct": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"mrp": {
						"type": "number",
						"format": "double"
					},
					"sellingPrice": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/mongoose.Types.ObjectId"
					},
					"sku": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/IcommonImage"
						},
						"type": "array"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"inStock": {
						"type": "boolean"
					},
					"brand": {
						"type": "string"
					},
					"unit": {
						"$ref": "#/components/schemas/ValidUnit"
					},
					"productType": {
						"$ref": "#/components/schemas/ProductType"
					},
					"productDetails": {
						"$ref": "#/components/schemas/ProductDetails"
					},
					"averageRating": {
						"type": "number",
						"format": "double"
					},
					"reviewCount": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"name",
					"slug",
					"mrp",
					"sellingPrice",
					"category",
					"sku",
					"quantity",
					"isActive",
					"isFeatured",
					"inStock",
					"unit",
					"productType",
					"productDetails",
					"averageRating",
					"reviewCount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponse_IProduct-or-null_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IProduct"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NullSuccessResponse": {
				"$ref": "#/components/schemas/SuccessResponse_null_"
			},
...
			"CustomSuccessResponse____": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {},
						"type": "object"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_IProduct_": {
				"properties": {
					"docs": {
						"items": {
							"$ref": "#/components/schemas/IProduct"
						},
						"type": "array"
					},
					"totalDocs": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPrevPage": {
						"type": "boolean"
					},
					"nextPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"prevPage": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"docs",
					"totalDocs",
					"limit",
					"page",
					"totalPages",
					"hasNextPage",
					"hasPrevPage",
					"nextPage",
					"prevPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomSuccessResponse_PaginatedResponse_IProduct__": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/PaginatedResponse_IProduct_"
					}
				},
				"required": [
					"success",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Enter JWT token with `Bearer <token>`"
			}
		}
	},
	"info": {
		"title": "milqit-ecom-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/customer/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_PaginatedResponse_CustomerProductResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all available products for customers\nOnly returns active and in-stock products",
				"summary": "Get Products for Customers",
				"tags": [
					"CUSTOMER: Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "inStock",
						"required": false,
						"schema": {
							"type": "boolean",
							"enum": [
								true
							]
						}
					},
					{
						"in": "query",
						"name": "brand",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "unit",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"piece",
								"kg",
								"gm",
								"litre",
								"ml",
								"pack",
								"dozen"
							]
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"name",
								"createdAt",
								"quantity",
								"unit",
								"price",
								"popularity",
								"rating",
								"relevance"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "isPublic",
						"required": false,
						"schema": {
							"type": "boolean",
							"enum": [
								true
							]
						}
					}
				]
			}
		},
		"/customer/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse_IProduct-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get a single product by ID for customers",
				"summary": "Get Product Details",
				"tags": [
					"CUSTOMER: Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/admin/products": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomSuccessResponse____"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Fresh Organic Milk",
											"description": "Premium quality organic milk from local farms",
											"mrp": 100,
											"sellingPrice": 85,
											"unit": "litre",
											"category": "60f7b3b3b3f1b40015c8e8a1",
											"quantity": 50,
											"productType": "Food",
											"brand": "FreshFarm",
											"sku": "FF-MILK-001",
											"isActive": true,
											"isFeatured": false
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Validation Failed"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new product with file uploads.\nNote: When uploading files, data must be sent as form fields.",
				"tags": [
					"ADMIN: Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"mrp": {
										"type": "string"
									},
									"sellingPrice": {
										"type": "string"
									},
									"unit": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"productType": {
										"type": "string"
									},
									"brand": {
										"type": "string"
									},
									"sku": {
										"type": "string"
									},
									"isActive": {
										"type": "string"
									},
									"isFeatured": {
										"type": "string"
									},
									"productDetails": {
										"type": "string"
									},
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"name",
									"description",
									"mrp",
									"sellingPrice",
									"unit",
									"category",
									"quantity",
									"productType"
								]
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomSuccessResponse_PaginatedResponse_IProduct__"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"ADMIN: Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "categoryId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "brand",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "unit",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"piece",
								"kg",
								"gm",
								"litre",
								"ml",
								"pack",
								"dozen"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"name",
								"price",
								"createdAt",
								"popularity",
								"rating",
								"quantity",
								"unit",
								"relevance"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean",
							"enum": [
								true
							]
						}
					},
					{
						"in": "query",
						"name": "isPublic",
						"required": false,
						"schema": {
							"type": "boolean",
							"enum": [
								true
							]
						}
					},
					{
						"in": "query",
						"name": "inStock",
						"required": false,
						"schema": {
							"type": "boolean",
							"enum": [
								true
							]
						}
					}
				]
			}
		},
		"/admin/products/{id}": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NullSuccessResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get a single product by its ID.",
				"summary": "Authored by MarotiKathoke at 2025-09-01 10:32:19",
				"tags": [
					"ADMIN: Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:5001/v1"
		}
	]
}